[feat] 무드차트 화면 오류 수정

+++ app/src/main/java/com/example/remind/data/model/response/GetFeelingPercentResponse.kt
@@ -7,5 +7,5 @@ data class GetFeelingPercentResponse(
 )
 data class PercentList(
     val feelingType: String,
-    val percent: Int
+    val percent: Double
 )
